ScriptHistory ScriptHistory gives us the ability to store portions of our script and navigate convos. WIP but handles very basic prompts. Highly referenced from OrionAlpha atm

+++ src/main/kotlin/script/ScriptMessageType.kt
@@ -1,6 +1,6 @@
 package script
 
-enum class ScriptMessageType (val type: Byte) {
+enum class ScriptMessageType (val type: Int) {
    Say(0),
    AskYesNo(1),
    AskText(2),
